feat: multiples functions

+++ fooddelivery/Controllers/Api/AddressesController.cs
@@ -7,8 +7,8 @@ using Microsoft.AspNetCore.Mvc;
 namespace fooddelivery.Controllers.Api
 {
     [ApiController]
-    [Route("api/[controller]")]
-    public class AddressesController:ControllerBase
+    [Route(("api/[controller]"))]
+    public class AddressesController : ControllerBase
     {
         private readonly IAddressService _addressService;
 
@@ -18,7 +18,7 @@ namespace fooddelivery.Controllers.Api
         }
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> Get(long id)
+        public async Task<IActionResult> Get(ulong id)
         {
             var result = await _addressService.GetByKeyAsync(id);
             return Ok(result);
@@ -27,24 +27,51 @@ namespace fooddelivery.Controllers.Api
         [HttpGet]
         public async Task<IActionResult> GetAll([FromQuery] AppView appview)
         {
-            var results = await _addressService.GetAllAsync(appview, x => x.City.Contains(appview.Search));
+            var results = await _addressService.GetAllAsync(appview, x => x.City.Contains(appview.Search)
+                                                                       || x.Neighborhood.Contains(appview.Search)
+                                                                       || x.State.Contains(appview.Search));
+
             return Ok(results);
         }
+
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] Address address)
         {
+            if (address == null)
+                return BadRequest();
+            if (!ModelState.IsValid)
+                return UnprocessableEntity(ModelState);
+
             await _addressService.AddAsync(address);
             return Ok(address);
         }
         [HttpDelete]
-        public async Task<IActionResult> Delete([FromRoute] long id)
+        public async Task<IActionResult> Delete([FromQuery] ulong id)
         {
+            var obj = _addressService.GetByKeyAsync(id);
+            if (obj == null)
+                return NotFound("recurso não encontrado");
+
             await _addressService.DeleteAsync(id);
             return Ok($"codigo {id} removido");
         }
-        [HttpPut]
-        public async Task<IActionResult> Update([FromBody] Address address)
+
+        [HttpPut("{id}")]
+        public async Task<IActionResult> Update(ulong id, [FromBody] Address address)
         {
+
+            var obj = _addressService.GetByKeyAsync(id);
+
+            if (obj == null)
+                return NotFound();
+
+            if (address == null)
+                return BadRequest();
+
+            if (!ModelState.IsValid)
+                return UnprocessableEntity(ModelState);
+
+            
             await _addressService.UpdateAsync(address);
             return Ok(address);
         }
