+++ fooddelivery/Controllers/Api/CategoriesController.cs
@@ -1,7 +1,6 @@
-using System.Linq;
 using System.Threading.Tasks;
-using fooddelivery.Database;
 using fooddelivery.Models;
+using fooddelivery.Models.Constants;
 using fooddelivery.Models.Helpers;
 using fooddelivery.Service.Interfaces;
 using Microsoft.AspNetCore.Authorization;
@@ -22,14 +21,14 @@ namespace fooddelivery.Controllers.Api
         }
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> Get(long id)
+        public async Task<IActionResult> Get(ulong id)
         {
             var result = await _categoryService.GetByKeyAsync(id);
             return Ok(result);
         }
 
-        [HttpGet]
-        public async Task<IActionResult> GetAll([FromRoute] AppView appview)
+        [HttpGet/*, Authorize(Policy = Policy.EmailVerified)*/]
+        public async Task<IActionResult> GetAll([FromQuery] AppView appview)
         {
             var results = await _categoryService.GetAllAsync(appview, x => x.Name.Contains(appview.Search));
             return Ok(results);
@@ -37,18 +36,40 @@ namespace fooddelivery.Controllers.Api
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] Category category)
         {
+            if (category == null)
+                return BadRequest();
+            if (!ModelState.IsValid)
+                return UnprocessableEntity(ModelState);
+
             await _categoryService.AddAsync(category);
             return Ok(category);
         }
         [HttpDelete]
-        public async Task<IActionResult> Delete([FromRoute] long id)
+        public async Task<IActionResult> Delete([FromQuery] ulong id)
         {
+
+            var obj = _categoryService.GetByKeyAsync(id);
+            if (obj == null)
+                return NotFound("recurso não encontrado");
+
+
             await _categoryService.DeleteAsync(id);
             return Ok($"codigo {id} removido");
         }
-        [HttpPut]
-        public async Task<IActionResult> Update([FromBody] Category category)
+        [HttpPut ("{id}")]
+        public async Task<IActionResult> Update(ulong id, [FromBody] Category category)
         {
+            var obj = _categoryService.GetByKeyAsync(id);
+
+            if (obj == null)
+                return NotFound();
+
+            if (category == null)
+                return BadRequest();
+
+            if (!ModelState.IsValid)
+                return UnprocessableEntity(ModelState);
+
             await _categoryService.UpdateAsync(category);
             return Ok(category);
         }
