+++ fooddelivery/Controllers/Api/FoodsController.cs
@@ -1,7 +1,9 @@
 using System.Threading.Tasks;
 using fooddelivery.Models;
+using fooddelivery.Models.Constants;
 using fooddelivery.Models.Helpers;
 using fooddelivery.Service.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace fooddelivery.Controllers.Api
@@ -18,21 +20,21 @@ namespace fooddelivery.Controllers.Api
         }
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> Get(long id)
+        public async Task<IActionResult> Get(ulong id)
         {
             var result = await _foodService.GetByKeyAsync(id);
             return Ok(result);
         }
-        
-        [HttpGet("ByCategoryId")]
-        public async Task<IActionResult> GetByCategoryId([FromRoute]long CategoryId, [FromRoute] AppView appview)
+
+        [HttpGet("ByCategoryId/{categoryid}")]
+        public async Task<IActionResult> GetByCategoryId([FromQuery] ulong CategoryId, [FromQuery] AppView appview)
         {
             var results = await _foodService.GetByCategoryIdAsync(CategoryId, appview);
             return Ok(results);
         }
 
-        [HttpGet]
-        public async Task<IActionResult> GetAll([FromRoute] AppView appview)
+        [HttpGet/*, Authorize(Policy = Policy.EmailVerified)*/]
+        public async Task<IActionResult> GetAll([FromQuery] AppView appview)
         {
             var results = await _foodService.GetAllAsync(appview, x => x.Name.Contains(appview.Search));
             return Ok(results);
@@ -40,18 +42,40 @@ namespace fooddelivery.Controllers.Api
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] Food food)
         {
+            if (food == null)
+                return BadRequest();
+            if (!ModelState.IsValid)
+                return UnprocessableEntity(ModelState);
+
             await _foodService.AddAsync(food);
             return Ok(food);
         }
         [HttpDelete]
-        public async Task<IActionResult> Delete([FromRoute] long id)
+        public async Task<IActionResult> Delete([FromQuery] ulong id)
         {
+            var obj = _foodService.GetByKeyAsync(id);
+            if (obj == null)
+                return NotFound("recurso não encontrado");
+
+
             await _foodService.DeleteAsync(id);
             return Ok($"codigo {id} removido");
         }
-        [HttpPut]
-        public async Task<IActionResult> Update([FromBody] Food food)
+        
+        [HttpPut("{id}")]
+        public async Task<IActionResult> Update(ulong id, [FromBody] Food food)
         {
+            var obj = _foodService.GetByKeyAsync(id);
+
+            if (obj == null)
+                return NotFound();
+
+            if (food == null)
+                return BadRequest();
+
+            if (!ModelState.IsValid)
+                return UnprocessableEntity(ModelState);
+               
             await _foodService.UpdateAsync(food);
             return Ok(food);
         }
