+++ fooddelivery/Controllers/Api/ImagesController.cs
@@ -9,7 +9,8 @@ namespace fooddelivery.Controllers.Api
     [ApiController]
     [Route("api/[controller]")]
     public class ImagesController : ControllerBase
-    {        private readonly IImageService _imageService;
+    {
+        private readonly IImageService _imageService;
 
         public ImagesController(IImageService imageService)
         {
@@ -17,14 +18,14 @@ namespace fooddelivery.Controllers.Api
         }
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> Get(long id)
+        public async Task<IActionResult> Get(ulong id)
         {
             var result = await _imageService.GetByKeyAsync(id);
             return Ok(result);
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll([FromRoute] AppView appview)
+        public async Task<IActionResult> GetAll([FromQuery] AppView appview)
         {
             var results = await _imageService.GetAllAsync(appview, x => x.Name.Contains(appview.Search));
             return Ok(results);
@@ -32,18 +33,37 @@ namespace fooddelivery.Controllers.Api
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] Image image)
         {
+            if (image == null)
+                return BadRequest();
+            if (!ModelState.IsValid)
+                return UnprocessableEntity(ModelState);
+
             await _imageService.AddAsync(image);
             return Ok(image);
         }
         [HttpDelete]
-        public async Task<IActionResult> Delete([FromRoute] long id)
+        public async Task<IActionResult> Delete([FromQuery] ulong id)
         {
+            var obj = _imageService.GetByKeyAsync(id);
+            if (obj == null)
+                return NotFound("recurso não encontrado");
+
             await _imageService.DeleteAsync(id);
             return Ok($"codigo {id} removido");
         }
-        [HttpPut]
-        public async Task<IActionResult> Update([FromBody] Image image)
+        [HttpPut("{id}")]
+        public async Task<IActionResult> Update(ulong id, [FromBody] Image image)
         {
+            var obj = _imageService.GetByKeyAsync(id);
+
+            if (obj == null)
+                return NotFound();
+
+            if (image == null)
+                return BadRequest();
+
+            if (!ModelState.IsValid)
+                return UnprocessableEntity(ModelState);
             await _imageService.UpdateAsync(image);
             return Ok(image);
         }
