+++ fooddelivery/Controllers/Api/SubordersController.cs
@@ -18,14 +18,14 @@ namespace fooddelivery.Controllers.Api
         }
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> Get(long id)
+        public async Task<IActionResult> Get(ulong id)
         {
             var result = await _suborderService.GetByKeyAsync(id);
             return Ok(result);
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll([FromRoute] AppView appview)
+        public async Task<IActionResult> GetAll([FromQuery] AppView appview)
         {
             var results = await _suborderService.GetAllAsync(appview, null);
             return Ok(results);
@@ -33,18 +33,37 @@ namespace fooddelivery.Controllers.Api
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] Suborder suborder)
         {
+            if (suborder == null)
+                return BadRequest();
+            if (!ModelState.IsValid)
+                return UnprocessableEntity(ModelState);
+
             await _suborderService.AddAsync(suborder);
             return Ok(suborder);
         }
         [HttpDelete]
-        public async Task<IActionResult> Delete([FromRoute] long id)
+        public async Task<IActionResult> Delete([FromQuery] ulong id)
         {
+            var obj = _suborderService.GetByKeyAsync(id);
+            if (obj == null)
+                return NotFound("recurso não encontrado");
+
             await _suborderService.DeleteAsync(id);
             return Ok($"codigo {id} removido");
         }
-        [HttpPut]
-        public async Task<IActionResult> Update([FromBody] Suborder suborder)
+        [HttpPut("{id}")]
+        public async Task<IActionResult> Update(ulong id, [FromBody] Suborder suborder)
         {
+            var obj = _suborderService.GetByKeyAsync(id);
+
+            if (obj == null)
+                return NotFound();
+
+            if (suborder == null)
+                return BadRequest();
+
+            if (!ModelState.IsValid)
+                return UnprocessableEntity(ModelState);
             await _suborderService.UpdateAsync(suborder);
             return Ok(suborder);
         }
